pyproject.toml
[project]
name = "quantforge"
version = "0.1.0"
description = "A modern quantitative trading system for A-share market"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "akshare>=1.12",
    "aiohttp>=3.9",
    "click>=8.1",
    "rich>=13.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "torch>=2.1",
]
api = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "anthropic>=0.25",
    "openai>=1.0",
    "python-multipart>=0.0.9",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]
all = ["quantforge[ml,api,dev]"]

[project.scripts]
quantforge = "quantforge.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
